+++ src/lib/remote_client.c
@@ -131,7 +131,8 @@ void remote_read_response(struct remote_handle *hndle,char **buf,int *length)
   memmove(*buf,idx,*length);
 };
 
-void remote_read_data(struct remote_handle *hndle,long long unsigned int offset, char **data, unsigned int *length)
+void remote_read_data(struct remote_handle *hndle,uint64_t offset, char **data, 
+		      uint32_t *length)
 {
   char tmp[BUFSIZE];
   int len;
@@ -147,7 +148,7 @@ void remote_read_data(struct remote_handle *hndle,long long unsigned int offset,
 
 void test_harness(struct remote_handle *hndle,char *filename) {
   int fd1;
-  long long unsigned int offset;
+  uint64_t offset;
   int count=0;
   int read_size;
   char *data1,*data2;
